changed "double" -> "FLOAT"

+++ libmp3lame/takehiro.c
@@ -72,8 +72,8 @@ static int quantize_xrpow(const FLOAT *xp, gr_info *gi)
 	sfb++;
 	do {
 #ifdef TAKEHIRO_IEEE754_HACK
-	    double x0 = istep * xp[0] + MAGIC_FLOAT;
-	    double x1 = istep * xp[1] + MAGIC_FLOAT;
+	    FLOAT x0 = istep * xp[0] + MAGIC_FLOAT;
+	    FLOAT x1 = istep * xp[1] + MAGIC_FLOAT;
 	    xp += 2;
 
 	    fi[0].f = x0;
